+++ cmd/proxy/cmd/proxy.go
@@ -1,4 +1,5 @@
 // Copyright 2015 Sorint.lab
+// Copyright 2026 WoozyMasta
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,6 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Package cmd implements the stolon-proxy command.
 package cmd
 
 import (
@@ -29,16 +31,17 @@ import (
 	"github.com/sorintlab/stolon/internal/flagutil"
 	slog "github.com/sorintlab/stolon/internal/log"
 	"github.com/sorintlab/stolon/internal/store"
+	"github.com/sorintlab/stolon/internal/tcpproxy"
 	"github.com/sorintlab/stolon/internal/util"
 
 	"github.com/davecgh/go-spew/spew"
-	"github.com/sorintlab/pollon"
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
 )
 
 var log = slog.S()
 
+// CmdProxy is the root command for stolon-proxy.
 var CmdProxy = &cobra.Command{
 	Use:     "stolon-proxy",
 	Run:     proxy,
@@ -76,6 +79,7 @@ func init() {
 	}
 }
 
+// ClusterChecker keeps the local TCP proxy aligned with cluster data.
 type ClusterChecker struct {
 	uid           string
 	listenAddress string
@@ -83,18 +87,18 @@ type ClusterChecker struct {
 
 	stopListening bool
 
-	listener         *net.TCPListener
-	pp               *pollon.Proxy
-	e                store.Store
-	endPollonProxyCh chan error
+	tcpProxy      *tcpproxy.Proxy
+	e             store.Store
+	endTCPProxyCh chan error
 
-	pollonMutex sync.Mutex
+	tcpProxyMutex sync.Mutex
 
 	proxyCheckInterval time.Duration
 	proxyTimeout       time.Duration
 	configMutex        sync.Mutex
 }
 
+// NewClusterChecker creates a ClusterChecker from proxy configuration.
 func NewClusterChecker(uid string, cfg config) (*ClusterChecker, error) {
 	e, err := cmd.NewStore(&cfg.CommonConfig)
 	if err != nil {
@@ -102,29 +106,30 @@ func NewClusterChecker(uid string, cfg config) (*ClusterChecker, error) {
 	}
 
 	return &ClusterChecker{
-		uid:              uid,
-		listenAddress:    cfg.listenAddress,
-		port:             cfg.port,
-		stopListening:    cfg.stopListening,
-		e:                e,
-		endPollonProxyCh: make(chan error),
+		uid:           uid,
+		listenAddress: cfg.listenAddress,
+		port:          cfg.port,
+		stopListening: cfg.stopListening,
+		e:             e,
+		endTCPProxyCh: make(chan error),
 
 		proxyCheckInterval: cluster.DefaultProxyCheckInterval,
 		proxyTimeout:       cluster.DefaultProxyTimeout,
 	}, nil
 }
 
-func (c *ClusterChecker) startPollonProxy() error {
-	c.pollonMutex.Lock()
-	defer c.pollonMutex.Unlock()
-	if c.pp != nil {
+func (c *ClusterChecker) startTCPProxy() error {
+	c.tcpProxyMutex.Lock()
+	defer c.tcpProxyMutex.Unlock()
+	if c.tcpProxy != nil {
 		return nil
 	}
 
 	log.Infow("Starting proxying")
-	addr, err := net.ResolveTCPAddr("tcp", net.JoinHostPort(cfg.listenAddress, cfg.port))
+	listenAddr := net.JoinHostPort(c.listenAddress, c.port)
+	addr, err := net.ResolveTCPAddr("tcp", listenAddr)
 	if err != nil {
-		return fmt.Errorf("error resolving tcp addr %q: %v", addr.String(), err)
+		return fmt.Errorf("error resolving tcp addr %q: %v", listenAddr, err)
 	}
 
 	listener, err := net.ListenTCP("tcp", addr)
@@ -132,46 +137,41 @@ func (c *ClusterChecker) startPollonProxy() error {
 		return fmt.Errorf("error listening on tcp addr %q: %v", addr.String(), err)
 	}
 
-	pp, err := pollon.NewProxy(listener)
-	if err != nil {
-		return fmt.Errorf("error creating pollon proxy: %v", err)
-	}
-	pp.SetKeepAlive(true)
-	pp.SetKeepAliveIdle(time.Duration(cfg.keepAliveIdle) * time.Second)
-	pp.SetKeepAliveCount(cfg.keepAliveCount)
-	pp.SetKeepAliveInterval(time.Duration(cfg.keepAliveInterval) * time.Second)
+	pp := tcpproxy.New(listener, tcpproxy.Options{
+		KeepAlive:         true,
+		KeepAliveIdle:     time.Duration(cfg.keepAliveIdle) * time.Second,
+		KeepAliveCount:    cfg.keepAliveCount,
+		KeepAliveInterval: time.Duration(cfg.keepAliveInterval) * time.Second,
+	})
 
-	c.pp = pp
-	c.listener = listener
-
+	c.tcpProxy = pp
 	go func() {
-		c.endPollonProxyCh <- c.pp.Start()
+		c.endTCPProxyCh <- pp.Start()
 	}()
 
 	return nil
 }
 
-func (c *ClusterChecker) stopPollonProxy() {
-	c.pollonMutex.Lock()
-	defer c.pollonMutex.Unlock()
-	if c.pp != nil {
+func (c *ClusterChecker) stopTCPProxy() {
+	c.tcpProxyMutex.Lock()
+	defer c.tcpProxyMutex.Unlock()
+	if c.tcpProxy != nil {
 		log.Infow("Stopping listening")
-		c.pp.Stop()
-		c.pp = nil
-		c.listener.Close()
-		c.listener = nil
+		c.tcpProxy.Stop()
+		c.tcpProxy = nil
 	}
 }
 
-func (c *ClusterChecker) sendPollonConfData(confData pollon.ConfData) {
-	c.pollonMutex.Lock()
-	defer c.pollonMutex.Unlock()
-	if c.pp != nil {
-		c.pp.C <- confData
+func (c *ClusterChecker) setProxyDestination(addr *net.TCPAddr) {
+	c.tcpProxyMutex.Lock()
+	defer c.tcpProxyMutex.Unlock()
+	if c.tcpProxy != nil {
+		c.tcpProxy.SetDestination(addr)
 	}
 }
 
-func (c *ClusterChecker) SetProxyInfo(e store.Store, generation int64, proxyTimeout time.Duration) error {
+// SetProxyInfo updates this proxy's liveness and generation information.
+func (c *ClusterChecker) SetProxyInfo(generation int64, proxyTimeout time.Duration) error {
 	proxyInfo := &cluster.ProxyInfo{
 		InfoUID:      common.UID(),
 		UID:          c.uid,
@@ -186,30 +186,30 @@ func (c *ClusterChecker) SetProxyInfo(e store.Store, generation int64, proxyTime
 	return nil
 }
 
-// Check reads the cluster data and applies the right pollon configuration.
+// Check reads the cluster data and applies the right proxy configuration.
 func (c *ClusterChecker) Check() error {
 	cd, _, err := c.e.GetClusterData(context.TODO())
 	if err != nil {
 		return fmt.Errorf("cannot get cluster data: %v", err)
 	}
 
-	// Start pollon if not active
-	if err = c.startPollonProxy(); err != nil {
+	// Start proxy if not active.
+	if err = c.startTCPProxy(); err != nil {
 		return fmt.Errorf("failed to start proxy: %v", err)
 	}
 
 	log.Debugf("cd dump: %s", spew.Sdump(cd))
 	if cd == nil {
 		log.Infow("no clusterdata available, closing connections to master")
-		c.sendPollonConfData(pollon.ConfData{DestAddr: nil})
+		c.setProxyDestination(nil)
 		return nil
 	}
 	if cd.FormatVersion != cluster.CurrentCDFormatVersion {
-		c.sendPollonConfData(pollon.ConfData{DestAddr: nil})
+		c.setProxyDestination(nil)
 		return fmt.Errorf("unsupported clusterdata format version: %d", cd.FormatVersion)
 	}
 	if err = cd.Cluster.Spec.Validate(); err != nil {
-		c.sendPollonConfData(pollon.ConfData{DestAddr: nil})
+		c.setProxyDestination(nil)
 		return fmt.Errorf("clusterdata validation failed: %v", err)
 	}
 
@@ -219,18 +219,15 @@ func (c *ClusterChecker) Check() error {
 	// use the greater between the current proxy timeout and the one defined in the cluster spec if they're different.
 	// in this way we're updating our proxyInfo using a timeout that is greater or equal the current active timeout timer.
 	c.configMutex.Lock()
-	proxyTimeout := c.proxyTimeout
-	if cdProxyTimeout > proxyTimeout {
-		proxyTimeout = cdProxyTimeout
-	}
+	proxyTimeout := max(c.proxyTimeout, cdProxyTimeout)
 	c.configMutex.Unlock()
 
 	proxy := cd.Proxy
 	if proxy == nil {
 		log.Infow("no proxy object available, closing connections to master")
-		c.sendPollonConfData(pollon.ConfData{DestAddr: nil})
+		c.setProxyDestination(nil)
 		// ignore errors on setting proxy info
-		if err = c.SetProxyInfo(c.e, cluster.NoGeneration, proxyTimeout); err != nil {
+		if err = c.SetProxyInfo(cluster.NoGeneration, proxyTimeout); err != nil {
 			log.Errorw("failed to update proxyInfo", zap.Error(err))
 		} else {
 			// update proxyCheckinterval and proxyTimeout only if we successfully updated our proxy info
@@ -245,9 +242,9 @@ func (c *ClusterChecker) Check() error {
 	db, ok := cd.DBs[proxy.Spec.MasterDBUID]
 	if !ok {
 		log.Infow("no db object available, closing connections to master", "db", proxy.Spec.MasterDBUID)
-		c.sendPollonConfData(pollon.ConfData{DestAddr: nil})
+		c.setProxyDestination(nil)
 		// ignore errors on setting proxy info
-		if err = c.SetProxyInfo(c.e, proxy.Generation, proxyTimeout); err != nil {
+		if err = c.SetProxyInfo(proxy.Generation, proxyTimeout); err != nil {
 			log.Errorw("failed to update proxyInfo", zap.Error(err))
 		} else {
 			// update proxyCheckinterval and proxyTimeout only if we successfully updated our proxy info
@@ -262,37 +259,37 @@ func (c *ClusterChecker) Check() error {
 	addr, err := net.ResolveTCPAddr("tcp", net.JoinHostPort(db.Status.ListenAddress, db.Status.Port))
 	if err != nil {
 		log.Errorw("cannot resolve db address", zap.Error(err))
-		c.sendPollonConfData(pollon.ConfData{DestAddr: nil})
+		c.setProxyDestination(nil)
 		return nil
 	}
 	log.Infow("master address", "address", addr)
-	if err = c.SetProxyInfo(c.e, proxy.Generation, proxyTimeout); err != nil {
+	if err = c.SetProxyInfo(proxy.Generation, proxyTimeout); err != nil {
 		// if we failed to update our proxy info when a master is defined we
 		// cannot ignore this error since the sentinel won't know that we exist
 		// and are sending connections to a master so, when electing a new
 		// master, it'll not wait for us to close connections to the old one.
 		return fmt.Errorf("failed to update proxyInfo: %v", err)
-	} else {
-		// update proxyCheckinterval and proxyTimeout only if we successfully updated our proxy info
-		c.configMutex.Lock()
-		c.proxyCheckInterval = cdProxyCheckInterval
-		c.proxyTimeout = cdProxyTimeout
-		c.configMutex.Unlock()
 	}
+	// update proxyCheckInterval and proxyTimeout only if we successfully updated our proxy info
+	c.configMutex.Lock()
+	c.proxyCheckInterval = cdProxyCheckInterval
+	c.proxyTimeout = cdProxyTimeout
+	c.configMutex.Unlock()
 
 	// start proxing only if we are inside enabledProxies, this ensures that the
 	// sentinel has read our proxyinfo and knows we are alive
 	if util.StringInSlice(proxy.Spec.EnabledProxies, c.uid) {
 		log.Infow("proxying to master address", "address", addr)
-		c.sendPollonConfData(pollon.ConfData{DestAddr: addr})
+		c.setProxyDestination(addr)
 	} else {
 		log.Infow("not proxying to master address since we aren't in the enabled proxies list", "address", addr)
-		c.sendPollonConfData(pollon.ConfData{DestAddr: nil})
+		c.setProxyDestination(nil)
 	}
 
 	return nil
 }
 
+// TimeoutChecker closes proxy connections when cluster checks stop succeeding.
 func (c *ClusterChecker) TimeoutChecker(checkOkCh chan struct{}) {
 	c.configMutex.Lock()
 	timeoutTimer := time.NewTimer(c.proxyTimeout)
@@ -305,9 +302,9 @@ func (c *ClusterChecker) TimeoutChecker(checkOkCh chan struct{}) {
 			// if the check timeouts close all connections and stop listening
 			// (for example to avoid load balancers forward connections to us
 			// since we aren't ready or in a bad state)
-			c.sendPollonConfData(pollon.ConfData{DestAddr: nil})
+			c.setProxyDestination(nil)
 			if c.stopListening {
-				c.stopPollonProxy()
+				c.stopTCPProxy()
 			}
 
 		case <-checkOkCh:
@@ -323,6 +320,7 @@ func (c *ClusterChecker) TimeoutChecker(checkOkCh chan struct{}) {
 	}
 }
 
+// Start runs the cluster checker loop.
 func (c *ClusterChecker) Start() error {
 	checkOkCh := make(chan struct{})
 	checkCh := make(chan error)
@@ -352,7 +350,7 @@ func (c *ClusterChecker) Start() error {
 			timerCh = time.NewTimer(c.proxyCheckInterval).C
 			c.configMutex.Unlock()
 
-		case err := <-c.endPollonProxyCh:
+		case err := <-c.endTCPProxyCh:
 			if err != nil {
 				return fmt.Errorf("proxy error: %v", err)
 			}
@@ -360,6 +358,7 @@ func (c *ClusterChecker) Start() error {
 	}
 }
 
+// Execute runs the stolon-proxy command.
 func Execute() {
 	if err := flagutil.SetFlagsFromEnv(CmdProxy.PersistentFlags(), "STPROXY"); err != nil {
 		log.Fatal(err)
@@ -370,7 +369,7 @@ func Execute() {
 	}
 }
 
-func proxy(c *cobra.Command, args []string) {
+func proxy(c *cobra.Command, _ []string) {
 	switch cfg.LogLevel {
 	case "error":
 		slog.SetLevel(zap.ErrorLevel)
@@ -392,10 +391,10 @@ func proxy(c *cobra.Command, args []string) {
 	if slog.IsDebug() {
 		if cmd.IsColorLoggerEnable(c, &cfg.CommonConfig) {
 			stdlog := slog.StdLogColor()
-			pollon.SetLogger(stdlog)
+			tcpproxy.SetLogger(stdlog)
 		} else {
 			stdlog := slog.StdLog()
-			pollon.SetLogger(stdlog)
+			tcpproxy.SetLogger(stdlog)
 		}
 	}
 
@@ -421,7 +420,11 @@ func proxy(c *cobra.Command, args []string) {
 	if cfg.MetricsListenAddress != "" {
 		http.Handle("/metrics", promhttp.Handler())
 		go func() {
-			err := http.ListenAndServe(cfg.MetricsListenAddress, nil)
+			server := &http.Server{
+				Addr:              cfg.MetricsListenAddress,
+				ReadHeaderTimeout: 5 * time.Second,
+			}
+			err := server.ListenAndServe()
 			if err != nil {
 				log.Fatalf("metrics http server error", zap.Error(err))
 			}